Committing a test that gets X/Y Vectors.
This probably won't compile (for you) because I haven't committed some of the
class changes in addition to this class (I still need to clean up some
code).

Backing up (committing) this file (test_vector_three.cpp) because this is the first
time that I've managed to map the "X" and "Y" directions in real life to
"X" and "Y" vectors that the Kinect sees.